fix: update messages, update modal component, add upload image component
