+++ apps/u3/src/components/common/modal/ModalBase.tsx
@@ -2,11 +2,10 @@
  * @Author: shixuewen friendlysxw@163.com
  * @Date: 2022-07-01 15:14:44
  * @LastEditors: shixuewen friendlysxw@163.com
- * @LastEditTime: 2023-01-31 19:20:51
+ * @LastEditTime: 2023-02-02 14:23:23
  * @Description: 基础按钮
  */
-import React from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import ReactModal from 'react-modal';
 import { isMobile } from 'react-device-detect';
 
@@ -18,11 +17,12 @@ const customStyles = {
     position: 'fixed',
     width: '100vw',
     height: '100vh',
-    overflow: 'auto',
     top: 0,
     left: 0,
     right: 0,
     bottom: 0,
+    padding: 0,
+    overflow: 'auto',
     backgroundColor: 'rgba(0, 0, 0, 0.5)',
     zIndex: 1,
   },
@@ -31,26 +31,27 @@ const customStyles = {
     height: 'fit-content',
     padding: 0,
     margin: '0 auto',
-    overflow: 'auto',
     WebkitOverflowScrolling: 'touch',
     border: 'none',
     background: 'none',
+    inset: '0px',
   },
 };
 function ModalBase({
   children,
   backdropFilter,
+  style,
   ...otherProps
 }: ModalBaseProps) {
-  const { overlay } = customStyles;
+  const { overlay, content } = customStyles;
   if (backdropFilter) {
     Object.assign(overlay, {
       backdropFilter: 'blur(12px)',
     });
   }
   const styles = {
-    overlay,
-    content: customStyles.content,
+    overlay: { ...overlay, ...(style?.overlay || {}) },
+    content: { ...content, ...(style?.content || {}) },
   };
   return (
     <ReactModal style={styles as ReactModal.Styles} {...otherProps}>
