+++ apps/u3/src/utils/message.ts
@@ -2,9 +2,10 @@
  * @Author: shixuewen friendlysxw@163.com
  * @Date: 2023-02-01 18:00:12
  * @LastEditors: shixuewen friendlysxw@163.com
- * @LastEditTime: 2023-02-01 18:56:59
+ * @LastEditTime: 2023-02-02 14:05:43
  * @Description: file description
  */
+import { EVENT_IMAGE_SIZE_LIMIT } from '../constants';
 import { CONTENT_ADMIN_PLUS_SCORE_STEP } from './content';
 import { EVENT_ADMIN_PLUS_SCORE_STEP } from './event';
 
@@ -32,5 +33,8 @@ export const messages = {
   },
   common: {
     error: 'Something seems wrong, please try again.',
+    upload_img: 'Upload successfully!',
+    upload_img_limit: `File Too Large, ${EVENT_IMAGE_SIZE_LIMIT / 1024}k limit`,
+    copy: 'Copied!',
   },
 };
